requirements.txt
aiogram==3.4.1
python-dotenv
requests
aiogram==3.4.1
python-dotenv
requests
aiogram>=3.4.1,<3.5
pydantic>=2.4.1,<2.6
aiohttp==3.9.5
python-dotenv==1.0.1
ics==0.7.2
